+++ CHANGELOG.md
@@ -4,9 +4,14 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 2.0.2 (TBD)
+## [2.0.2](https://github.com/ngageoint/geopackage-java/releases/tag/2.0.2) (03-20-2018)
 
-* TBD
+* Tile Scaling extension for generating missing tiles using nearby zoom levels
+* Skip tiles drawn from features when no features overlap the tile
+* Tile DAO approximate zoom level methods
+* Tile Generator fix to save updated bounds in the Tile Matrix Set
+* Tile Generator projection transformations only when projections differ
+* geopackage-core version updated to 2.0.2
 
 ## [2.0.1](https://github.com/ngageoint/geopackage-java/releases/tag/2.0.1) (02-13-2018)